To uninstall,let the user remove the file himself.
It is critical to delete files from a system and it could go very wrong.
After files have been installed on the target, CMake generates an
install_manifest.txt file.
Removing all files specified in this manifest is equivalent to an
uninstallation.

It would also avoid that we have such moments:
https://github.com/MrMEEE/bumblebee/commit/a047be85247755cdbe0acce6