Added libnacl support to build.
There is no easy way to integrate NaCL in the build enviroment as it is
lacking of Makefiles.
Its build script englobs header generation, self test, compiling and
linking in a single operation which cannot splitted in multiple smaller
steps.

As a workaround, the user will have separetely build NaCL for his target
and input the path toward the header files and the library.