build_nacl: 'hack' for fixing build with nacl for selecting right arch
NaCl tries to build for all possible architectures on the system, i.e. on
x86_64 it builds the lib for amd64 and for x86. On at least 2 standard
Debian-based x86_64 systems, the current build has selected the x86 library
version instead of amd64, hence the cmake build fails. This hack chooses
the right arch with the currently used gcc that is also used in the cmake
build.

Signed-off-by: Daniel Borkmann <borkmann@iogearbox.net>