scripts: update-oui: Add OUI list update script
Instead of having to maually update the oui.conf file used by
netsniff-ng to resolve OUIs to vendor names, provide a script which
fetches the necessary information from any file in the format as
provided by the IEEE at
http://standards.ieee.org/develop/regauth/oui/oui.txt

This is also the default URL the script will fetch the information from.

The script runs with Python 2.x and 3.x and only uses the respective
standard libraries.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>