curvetun: Added error reporting for mal-formed tcp/udp carrier types
I personally had mis-spelled udp (upd) and found myself diving into source
to determine why tcp was repeatedly being used instead of udp.

I'm uncertain whether the general lack of error checking and reporting
is intentional, or simply due to the very large task of getting the tools
up and running.

Further, if error checking is to be done on parameters and inputs, I'm
personally uneducated on the best method for reporting them.  I assumed
syslog(), as that is used frequently in setup/teardown messaging for
curvetun.  Is the inclusion of syslog.h reasonable here?