netsniff-ng trafgen: check return value of pull_and_flush_ring
Let us check the return value when the timer triggers a TX flush
request to the kernel. However, ignore the case of BADFS and NOBUFS.
The socket could already have been closed before the timer triggers
in the first case, and in the second, we just let the next timer
continue processing if currently the buffer space is exhausted.

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>