trafgen: fix bug in packet scheduling
When having a config like cpu(1): {...}, and one runs it with -n1, trafgen
behaves buggy, I also noted that in other situations. Fix this by letting
the loop also return on CPU state RES, and do not perform this stupid magic
in the main routine. So far it seems to work now after some basic tests I
did.

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>