pcap: let netsniff-ng also record pkt timestamp source
With commit [1] in the kernel, we can also store the timestamp
source in the pcap packet header for later analysis. We do this
by splitting the netsniff-ng's u32 ifindex into u16 tsource and
u16 ifindex. Older kernel do not support the timestamp source in
PF_PACKET, so it will stay 0 and is compatible with older
netsniff-ng binaries.

  [1] http://thread.gmane.org/gmane.linux.network/266878/

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>