netsniff-ng: v3: fix packet accounting on --num
We need to carry frame_count through multiple calls of walk function
to account correctly for --num <pkts>. Also, move socket stats printing
into rx ring, since it belongs there.

Todo: the kernel socket seems to have a different count that what we
see. This needs to be fixed one way or the other. Not yet sure what's
causing this.

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>