ifpps: Report more detailed memory stats and number of total processes
Instead of just giving the percentage of memory usage, report the total,
used, active and inactive memory values. Also report total, free and
cached for swap.

Instead of reporting the memory in the 'sys' line add two new lines for
these reports and use the free space to report the total number of
processes.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>