oui: Make parsing of oui.conf more robust
In the unlikely case of the oui.conf file not conforming to the required
format (e.g. by the user supplying one by himself), we would end up -
in the worst case dereferencing a null pointer, if the condiition
(ptr = strstr(buff, ", ")) is false. Thus make the entire parsing a bit
more robust and just ignore lines not following the required format.

The null pointer dereference was found by the Coverity scanner.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>