dissector: icmpv6: Fix possible null pointer dereferences
The Coverity scanner found several possible null pointer dereferences in
the ICMPv6 dissector. These are all related to not checking the return
value of pkt_pull(). Sometimes pkt_pull(()) is called iteratively based
on a length value in the encountered packet, so this could possibly be
hit in case an invalid packet is crafted accordingly.

Fix all by checking the return value of pkt_pull() consistently.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>