man: bpfc: add example of loadable x86_64 seccomp-BPF filter
Add the example in bpfc code from http://outflux.net/teach-seccomp/
to the man page, so that people can also experiment with seccomp
bpf and to demonstrate that there are no limits on bpfc when it
comes to this. Actually this code is even more efficient than the
one given in the example URL above.

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>