configure: Add configure script used to detect NaCl
In order to use an installation of NaCl provided by the distro or by
previous manual installation, provide a configure script which checks
for nacl's presence.

Also check of libnl-genl headers. More checks of the other libraries
depended on by netsniff-ng as well as checks for necessary kernel
headers should be added here too.

The configure script is preliminary and has not yet extensively been
tested on multiple systems.

The concept was inspired by the configure script of trinity and
iproute2.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>