Makefile, INSTALL: Simplify clean/uninstall Makefile targets
Streamline the Makfile targets by separating the clean targets (which
only affect the build tree) from the uninstall targets (affecting
files installed in the system).

Do this by renaming the 'distclean' target to 'uninstall' and removing
unnecessary aliases. The 'distclean' target serves now as a way to
remove build files and Config. The 'mrproper' will bring the source into
a pristine state as checked out from git.

Summary of the new/changed targets:

  clean:       remove all build related files (but not Config)
  distclean:   'make clean'  + remove Config
  mrproper:    remove all files not in git repository
  uninstall:   remove all files installed by us from the system, but not
               affecting the build environment.
  %_uninstall: tool specific uninstall

The following targets were removed:

  mostlyclean, realclean and clobber.

Also the target names in the INSTALL file are adjusted.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>