configure, Makefile: Conditionally build tools
Instead of halting out the entire build build if one dependency is not
available, only exclude the tools that link against it from building.
The configure script will now print a list of all tools that will be
built -- and in case some libaries are not available -- that will not be
built.

Also add checks for libcli and libnet which mausezahn depends on.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>