ifpps: align avg numbers and remove unit
In order to get a better overview, lets align the avg numbers to the
other shown decimals. However, for this we would need to remove the
avg units as we deal with floats here. But we know the units from the
above numbers anyway. I think this makes it more clear to follow.

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>