configure: redirect all compile output to our local config.log
On errors in the configuration, redirect everything to config.log,
and not some parts to config.log and some others to $TMPDIR/config.log.

This will give a nice summary of what goes wrong to the user. Also,
in case everything looks fine, the config.log file will be created,
but with size 0 and no warning is thrown. So looks good.

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>