build: add "+" version appendix automatically during build
For each build that has no tag name (thus for all commits except
releases), we add a "+" version appendix to indicate that it is
further than the rc tag itself, like in the kernel. Do this during
configure automatically. This is useful in case someone wants to
package a relase, so that the full version name indicates that.

Signed-off-by: Daniel Borkmann <dborkman@redhat.com>