build: Don't uninstall configuration files
Several of the tools install/uninstall files from $NCONF_FILES,
however if we call the uninstall target for a specific tool (e.g.  `make
astraceroute_uninstall') but other tools depending on one or more of the
config files are still installed, the file is still removed
unconditionally. As it would require quite some logic in the Makefile
in order to detect installed tools and their dependencies, we just don't
uninstall the configuration files anymore.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>