build: Add Makefile target for Coverity scanner
Add a Makefile target to build and upload a snapshot of netsniff-ng to
the Coverity scanner. The variables COV_EMAIL and COV_PASSWD need to be
set accordingly:

  make coverity COV_EMAIL=tklauser@distanz.ch COV_PASSWD=...

Currently the sub-make which is part of the cov-built call issues
warnings about undefined variables from the template. This can be
ignored, the build works just fine.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>