dissectors: ethernet: Handle multicast/broadcast addresses properly
Until now, we just looked up the vendor string based on the OUI for each
MAC address. Thus, multicast and broadcast addresses were just printed
as "Unknown" which is a bit misleading.

Improve this situation by checking bit 0 of the 1st octet of the address
and by checking for the broadcast address if it is set. If a
multicast/broadcast address is found, the respective string is returned.
In all other cases, the existing OUI lookup is done.

In the future we might extend this mechanism to look up well-known
multicast addresses.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>