build: add DESTDIR support
This change implements support for specifying an alternative
installation destination directory by setting DESTDIR when
calling 'make install'.

Example: make DESTDIR=/alternative-rootfs install

It is particularly useful when cross-compiling and installing
in a dedicated directory where for example the rootfs or a
build sandbox for the target architecture is mounted.

By using DESTDIR you can avoid to mess with PREFIX to include
the alternative installation directory!

Signed-off-by: Christian Wiese <chris@opensde.org>
Signed-off-by: Daniel Borkmann <dborkman@redhat.com>