conf: Use configuration file path from Makefile instead of hardcoded paths
Currently, the path to the configuration files is hardcoded to
$PREFIX/etc/netsniff-ng.

If we want to keep the configuration files outside of prefix (e.g.
during development), we need to pass the complete path as specified in
$ETCDIRE (and $ETCDIR respectively, which could e.g. be overriden on the
command line during build install) as a preprocessor define.

This way, we can e.g. install the configuration files to /etc while
installing the binaries in /usr/local with:

  $ make PREFIX=/usr/local ETCDIR=/etc
  $ make PREFIX=/usr/local ETCDIR=/etc install

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>