man: mausezahn: document rest of available config file options
This documents two more configuration file options that can be used.
One, automops, is in the development phase (not compiled by default),
and the other one, management-only, is usable. It prevents a user from
using a network interface to pass data traffic with the intention that
the interface is used for machine managment (remote control).

$ cat /etc/netsniff-ng/mausezahn.conf
user = mzadmin
password = better-password
management-only = eth0
automops = /path/to/an/automops-file.xml

Signed-off-by: Jon Schipp <jonschipp@gmail.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>