AUTHORS: Add Jon Schipp to major contributors list
Github shows 33 commits for me:
https://github.com/netsniff-ng/netsniff-ng/graphs/contributors

$ git shortlog | grep -i jon
Jon Schipp (29):
jonschipp (1):

Not sure what the difference is here, but this one will make 31 based
on git shortlog (I hope that's not cheating) :) Otherwise I can wait.

And I believe I have a few commits from the old history (is this still
up somewhere?). I think it used to be on the cryptoism site.

Signed-off-by: Jon Schipp <jschipp@illinois.edu>
Acked-by: Daniel Borkmann <borkmann@iogearbox.net>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>