build: Add build flag to enable sparse checking of C source files
Similar to the 'make C=1' for the Linux kernel, wire up the sparse
command for our build system in a similar fashion. Use 'make C=1 [tool]'
to check the source files using sparse. Of course, the sparse tool needs
to be installed and available in your $PATH for this to work.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>