netsniff-ng: Conditionally define POLLRDNORM where it is used
Instead of defining POLLRDNORM in built_in.h if it isn't provided by
system/kernel headers, define it in ring.h where its only user is. This
way we make sure that <poll.h> is included before checking #ifndef
POLLRDNORM. This fixes the following sparse warning:

/usr/include/bits/poll.h:32:10: warning: preprocessor token POLLRDNORM redefined
built_in.h:378:10: this was the original definition

Also remove the defines for POLLWRNORM and POLLRDHUP which are not used
anywhere in the code.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>