flowtop: Use integer conversion to determine PID proc entries
Try to convert the directory entry name to an unsigned integer with
strtoul() instead of using strspn() to determine if a proc entry is a
PID. If it is a valid PID (i.e. strtoul returned a value != 0), we can
directly use it to pass into walk_process() and there set
flow_entry->proc_num.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>