build: Don't overwrite CC variable
Don't reset the variable CC in the Makefile, but set it to gcc by
default (can be overriden when e.g. compiling with `make CC=clang').
This way we can also have non-verbose compiles with compilers other than
gcc. Furthermore make sure that the CC variable is also respected for
LD.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>