dissector: Don't re-initialize packet_types array
Both sparse and clang warn about the initializers overriding previous
initialization of the packet_types array. Since every access of the
packet_types array checks the value for NULL (the default value, since the
array is static) and prints a "?" if it isNULL, we don't need the prior
initialization with "?".

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>