build: Make ccache opt-in rather than opt-out
Using ccache by default for compilation if it is available on the system
might confuse people and in rare cases also lead to unexpected results.
Thus, from now on the use of ccache for compilation has to explicitely
specified by setting the CCACHE make variable to the name/path of the
ccache binary, i.e. `make CCACHE=ccache'

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>