netsniff-ng: Consider radiotap header of monitor dev
netsniff-ng does not check if monitor device includes radiotap
header which leads to the wrong 802.11 frame parsing.

Tested if the .pcap file is understandable by wireshark and if
dump info is basically correct, but did not test the case when xmit
packets from .pcap file to the output device and from the input device
to the output device.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>
[tklauser: whitespace changes]
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>