build: Remove DISTRO build flag, don't optimize by default
Currently, when building with DISTRO=0 we optimize for the build host's
architecture by default. This is not compatible with cross-compiling.
Remove the DISTRO flag and only build with -O2 and no -march/-mtune
flags by default.

Support for specifying optimization flags via configure script will be
added in a follow-up patch.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>