build: Restore support for cross-compiling the netsniff-ng toolkit
Ever since we switched to the hand-crafted ./configure script, support
for cross-compiling the netsniff-ng toolkit was basically broken.

Restore the abaility to cross-compile our tools by making ./configure
consider the CROSS_COMPILE and SYSROOT variables.

Example for cross-compiling on arm:

  $ CROSS_COMPILE=arm-linux-gnueabihf- \
    SYSROOT=/usr/arm-linux-gnueabihf \
    ./configure
  $ make

assuming the cross-compiled libraries (and their respective pkg-config
information) are in /usr/arm-linux-gnueabihf.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>