flowtop: Add connection traffic accounting
Mark each flow if it is visible on the screen to know if it is needed
update traffic acct info.

Changed to use non blocking recv of nf conntrack events to update
traffic accounting.

Now nf_conntrack is cloned when new flow entry is added to send dump
request which is used to update traffic accounting info (packet, bytes).

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>
[tklauser: Formatting changes]
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>