flowtop: Don't attempt to set sysctl values if initial read fails
Bail out early if we fail to read the current sysctl variable values for
net/netfilter/nf_conntrack_acct and net/netfilter/nf_conntrack_timestamp

Otherwise we'll not be able restore the previous value on exit/panic.
Moreover, if we fail to read the sysctl file, we usually also lack the
permissions to write it.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>