netsniff-ng: nlmsg: Display raw data if family is unknown
Currently we only support full dissection of RTNL netlink messages. For
non-RTNL message we only print the header and omit the data.

Change this behavior and print a full ascii/hex dump of the remaining
data (like it is done in dissector_entry_point() for trailing data after
all known protocols have been processed) to give the user a chance to
still inspect the message content.

Reported-by: Geoff Ladwig <gladwig@verdantnetworks.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>