netsniff-ng: Use time of SIGHUP time when rotating files prematurely
Use the timestamp of the SIGHUP in the file name when rotating file
prematurely instead of the file creation date, which might be delayed
depending on when the next packet arrives.

This should make it a bit easier to synchronize pcap files captures by
multiple instances of netsniff-ng on multiple interfaces.

Signed-off-by: Erik Bengtsson <e.bengtsson@gmail.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>