Allow IPv6 ranges to be specified for source and dest addresses.
This allows a user to pass a range of IPv6 addresses, either like:
fec0:5000::1-fec0:5000::100
or in CIDR notation:
fec0:5000::0/112
These can be used for the -A and/or -B command-line options.
The largest range that can be used is a /64.  In other words, if using CIDR
notation, the masklen must be <= 128 and >= 64.

Signed-off-by: Tommy Beadle <tbeadle@gmail.com>