netsniff-ng: Increment pkts_seen after packet type check
Currently in receive_to_xmit() pkts_seen is incremented before the
packet type check, but failing the latter will cause the packet to be
ignored, pretty much as if it failed to pass the filter.

This change moves the accunting after the check, as is currently
done in both walk_t3_block() and recv_only_or_dump().

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>