netsniff-ng: Allow to compile without libnl
One might not want to install libnl just for sniffing packets, for
example if netsniff-ng will be compiled on embedded or switch system.

Hide libnl dependend code if CONFIG_LIBNL=0.

In case the `--rfraw' option is used, the user will get a panic
message. In case of netlink messages being sniffed, they will not be
dissected.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>