trafgen: proto: Use field id as array index
Usually proto fields array are sorted in the same order as the
respective enum. Thus, the id can be used used as an array index for
faster lookup.

Add an explanatory comment and enforce the correspondence of id and
index using bug_on().

This will make csum field calculation a little faster at runtime.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>