flowtop: Render table row via raw ncurses buffer
Render each column to the ncurses raw buffer first, this buffer
contains ncurses {char:attr} elements which will be printed
to the screen after ui_table_row_show() will
be called (at the end of columns rendering by flowtop).

The reason of this change is to have easy way to
make horizontal scrolling over this buffer.

Approach is used from the 'htop' tool.

Signed-off-by: Vadim Kochan <vadim4j@gmail.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>