trafgen: fix dinc()/ddec() modifiers
currently, after dinc(), the valued stored inside the packet is
not in the (min, max) range but in the (0, max - min + 1) range,
'counter->val' should be used instead of 'val'.

Additionally the values computed for ddec() are corrupted, in:

val = (val - counter->inc) % (counter->min - counter->max + 1);

the divider is negative, we should use (counter->max - counter->min + 1)
as in the INC case.

Finally we can avoid the switch statement at update time, inverting
the value of 'counter->inc' for decrement and using a data type wide
enough for the 'inc' field.

v1 -> v2:
 - changed 'counter->inc' type to int

Signed-off-by: Paolo Abeni <pabeni@redhat.com>